requirements.txt
Appium-Python-Client
selenium
browsermob-proxy
pytest